# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CookPilot Backend API",
        "description": "\uc74c\uc131 \uae30\ubc18 \uc694\ub9ac \uc5b4\uc2dc\uc2a4\ud134\ud2b8 CookPilot \ubc31\uc5d4\ub4dc API. \ubaa8\ub4e0 \ub77c\uc6b0\ud2b8\ub294 /api/v1 \uc544\ub798\uc5d0 \uc788\uace0, \uc5d0\ub7ec\ub294 RFC-7807 ProblemDetail \ud615\uc2dd\uc774\ub2e4.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/recipes": {
            "get": {
                "tags": [
                    "recipe-controller"
                ],
                "operationId": "list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseRecipeSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recipes/{recipeId}": {
            "get": {
                "tags": [
                    "recipe-controller"
                ],
                "operationId": "get_1",
                "parameters": [
                    {
                        "name": "recipeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/recipes/search": {
            "get": {
                "tags": [
                    "recipe-controller"
                ],
                "operationId": "search",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "ingredient",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "cookingMethod",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "dishType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "hashtag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseRecipeSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PagedResponseRecipeSummaryResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeSummaryResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hasNext": {
                        "type": "boolean"
                    }
                }
            },
            "RecipeSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "hasPersonalVersion": {
                        "type": "boolean"
                    },
                    "latestPersonalVersionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "favorite": {
                        "type": "boolean"
                    },
                    "cookingMethod": {
                        "type": "string"
                    },
                    "dishType": {
                        "type": "string"
                    },
                    "hashtags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Recipe": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "baseServings": {
                        "type": "number",
                        "format": "double"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "cookingMethod": {
                        "type": "string"
                    },
                    "dishType": {
                        "type": "string"
                    },
                    "hashtags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeIngredient"
                        }
                    },
                    "steps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeStep"
                        }
                    }
                }
            },
            "RecipeIngredient": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "required": {
                        "type": "boolean"
                    }
                }
            },
            "RecipeStep": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "stepIndex": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "instruction": {
                        "type": "string"
                    },
                    "timerSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "cautionNote": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}